% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup}
\alias{cleanup}
\title{Cleanup}
\usage{
cleanup(df)
}
\arguments{
\item{df}{A data frame from qualtrics.}
}
\value{
A data frame
}
\description{
Removes the first two rows of the data frame df. Qualtrics data has two extra rows at the top that can mess up data manipulation.
}
