% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convertDataByName}
\alias{convertDataByName}
\title{Converts data based on the type markers in the column name.}
\usage{
convertDataByName(
  df,
  markers = list(numeric = "_num", logical = "_log", integer = "_int", complex =
    "_com", character = "_char", factor = "_fac")
)
}
\arguments{
\item{df}{A data frame from qualtrics.}

\item{markers}{TODO}
}
\value{
A data frame
}
\description{
Q
}
