% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convertDataByOrg}
\alias{convertDataByOrg}
\title{Converts data based on the types in the org data frame}
\usage{
convertDataByOrg(df, org)
}
\arguments{
\item{df}{A data frame from qualtrics.}

\item{org}{The organizational data frame.}
}
\value{
A data frame
}
\description{
Q
}
