% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAxisLabels.R
\name{getAxisLabels}
\alias{getAxisLabels}
\title{T}
\usage{
getAxisLabels(
  text_df,
  question_name,
  label_df = NULL,
  default = c("-3", "-2", "-1", "0", "1", "2", "3")
)
}
\arguments{
\item{text_df}{TODO}

\item{question_name}{TODO}

\item{label_df}{TODO}

\item{default}{TODO}
}
\value{
A data frame
}
\description{
Q
}
