% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLabel.R
\name{getLabel}
\alias{getLabel}
\title{T}
\usage{
getLabel(
  group_name,
  subgroup_name,
  linesize = 100,
  org = NULL,
  raw_df = NULL,
  label_df = NULL
)
}
\arguments{
\item{group_name}{TODO}

\item{subgroup_name}{TODO}

\item{linesize}{TODO}

\item{org}{TODO}

\item{raw_df}{TODO}

\item{label_df}{TODO}
}
\value{
A data frame
}
\description{
Q
}
