% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLabel.R
\name{getLabelByGrouping}
\alias{getLabelByGrouping}
\title{Retrieves the label for a graph.}
\usage{
getLabelByGrouping(labels, group_name, subgroup_name, linesize = 100, ...)
}
\arguments{
\item{labels}{TODO}

\item{group_name}{TODO}

\item{subgroup_name}{TODO}

\item{linesize}{When should we wrap our label to a new line?}
}
\value{
A data frame
}
\description{
getLabelByGrouping returns the label for a graph via a data frame that matches each group and subgroup to a string. The data frame has 3 columns minimum - subgroup, group and label.
}
