% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColumnNameByOrg.R
\name{getQuestionName}
\alias{getQuestionName}
\title{T}
\usage{
getQuestionName(org, group_name, subgroup_name)
}
\arguments{
\item{org}{TODO}

\item{group_name}{TODO}

\item{subgroup_name}{TODO}
}
\value{
A data frame
}
\description{
Q
}
