% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSummTable.R
\name{getSummTables}
\alias{getSummTables}
\title{T}
\usage{
getSummTables(df, org, group_name, width = 9, height = 6, unit = "in")
}
\arguments{
\item{df}{A data frame from qualtrics.}

\item{org}{TODO}

\item{group_name}{TODO}

\item{width}{TODO}

\item{height}{TODO}

\item{unit}{TODO}
}
\value{
A data frame
}
\description{
Q
}
