% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupData.R
\name{groupData}
\alias{groupData}
\title{Collects the data from the specified group}
\usage{
groupData(df, org, group_name, ignore_case = TRUE)
}
\arguments{
\item{df}{A data frame from qualtrics.}

\item{org}{The organizational data frame.}

\item{group_name}{A string containing the name of the group to gather data in.}

\item{ignore_case}{A boolean signifying if we should ignore case with group names, column names, and subgroup names. Default is TRUE.}
}
\value{
A data frame
}
\description{
TODO
}
