% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePage.R
\name{makePage}
\alias{makePage}
\title{Creates and saves a file of graphs, tables, and text}
\usage{
makePage(
  plot_list,
  page_title,
  table_list = NULL,
  table_height = 6,
  width = 9,
  plot_height = 4,
  unit = "in",
  columns = 1,
  padding = 0.75
)
}
\arguments{
\item{plot_list}{A list of plots.}

\item{page_title}{A string containing the title of page. This will be used in generating the file name.}

\item{width}{The width of the page}

\item{plot_height}{The height of each plot}

\item{unit}{The units the preceding measurements are in}

\item{columns}{How many columns the plots should be arranged in.}

\item{table_grobs}{An optional list of table grobs to place at the top of the page. We suggest tables containing the mean/standard deviation.}

\item{tables_height}{The height of the tables part of the page in total}
}
\value{
A data frame
}
\description{
Q
}
