% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGraphs.R
\name{makeGraphs}
\alias{makeGraphs}
\title{T}
\usage{
makeGraphs(
  df,
  org,
  group_name,
  usePercent = TRUE,
  getXAxisLimits = getAxisLimits,
  getXAxisLabels = getAxisLabels,
  getGraphLabels = getLabelByGrouping,
  text_df = NULL,
  graph_label_df = NULL
)
}
\arguments{
\item{label_df}{}
}
\value{

}
\description{
Assumes df has been cleaned and converted.
}
